requirements.txt
torch==2.3.0
torchvision==0.18.0
pyyaml==6.0.1